Fix support for Boost 1.87.0
Boost.Asio removed a bunch of long-deprecated stuff around
`boost::asio::io_service` and `boost::asio::ip::tcp::resolver::query`,
so we have to adapt it to the new format.

Note that I this isn't super tested, but it's not a very complex change,
and all tests seem to pass after this.